Add remove resource for identifier.
