+++ Sources/Cache/Database.swift
@@ -11,16 +11,26 @@ final class Database<Item> {
     }
     
     func resource(for identifier: UUID) -> ResourceItem? {
+        
         let predicate: (ResourceItem) -> Bool = {
             $0.identifier == identifier
         }
+                
         guard let resource = storage.first(where: predicate) else {
             return nil
         }
-        if resource.isExpired {
+        
+        guard !resource.isExpired else {
             storage.remove(resource)
             return nil
         }
+        
         return resource
+    }
+    
+    func removeResource(for identifier: UUID) {
+        if let resource = resource(for: identifier) {
+            storage.remove(resource)
+        }
     }
 }
